Добавляет контроллер для рендера задач
